Make local run independent of electron-forge
It's time to move to electron-builder, so we can properly
update Electron (primarily so we can support new ARM Macs,
but lots of other good reasons too). Electron Forge has
been stuck on an unsupported v5 vs an experimental
unreleased v6 for a long time, so it's time for something
properly supported.