Update Sentry
Here we update to a version that better supports Electron 11
(e.g. using compressed errors to avoid warnings) and also
update the DSN to the new format. Some errors currently seem
to not be being reported, hoping this might help with/fix
that en route.

This also removes the 'disable Sentry after quit' workaround. For
a few reasons:
- To make sure that's not hiding other issues somewhere.
- Because we've updated Electron, so hopefully much of the after-quit
  error noise that appeared before will be fixed now.
- Because it's not clear that it worked for native crashes, which
  were the main problem anyway.