Update server for Docker GA & custom CA trust
This enables Docker by default in the server, so that Docker is
automatically intercepted and handled in all terminals. The UI flag
still needs to be removed to allow attaching in the UI and tell users
that this is available, that's coming next if all goes well.

This also includes Java fixes, adds some extra ADB debugging info and
error handling, moves ephemeral servers onto ephemeral ports, includes
lots of minor dependency updates, fixes a warning when intercepting
Stripe in JS, and allows configuration of custom CA certificates, to
precisely configure which CAs are trusted in upstream connections.