Update server for assorted small bugfixes
Includes a fix to workaround node-gyp bug that includes build-tool
symlinks in the server node_modules, which is rejected by Mac code
signing. I'm going to fix this in the server too, and hopefully it'll
get fixed in node-gyp eventually as well, but this quick fix should let
us use the existing server build regardless.

Once this is fixed somewhere upstream, we can drop this workaround here.