Use crypto random bytes instead of a UUID for auth tokens
Both should be equally secure, but UUID formatting is needlessly long
and weird, it's an unnecessary dependency, and inefficient (it's 128
bits in theory, but 4 bites are constant so really just 122, and then
we're including dashes for formatting as well, and hex is not efficient
either).

Moving to crypto+base64url lets us boost the key size (122 to 160 bit)
and simultaneously shorten the token length. This also makes it clearer
that we're using cryptographically secure random values (really it
should been the same underlying source, but here we avoid hiding that
away & trusting the uuid dep).